Revert "PK Long->Integer로 변경"
This reverts commit 6da666ec8652891b5a8d64af023d1560f8c9f78f.

+++ src/main/java/kr/ac/bc/comon/springboot/common/repository/HistoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HistoryRepository extends JpaRepository<HistoryEntity, Integer> {
+public interface HistoryRepository extends JpaRepository<HistoryEntity, Long> {
 }
